server-py/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
pyyaml>=6.0.1
openai>=1.0.0
